separated spec files for button
