+++ cypress/integration/v2/blocks/button/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Button Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
